mainbuilder: Only clean up inactive_packages when there are files under them
git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@8445 b52b6941-3400-464c-9f42-43200397181c

+++ rebeccablackos_builder.sh
@@ -767,7 +767,10 @@ fi
 
 #Make a list of all current items under srcbuild
 mkdir -p "$BUILDLOCATION"/build/"$BUILDARCH"/srcbuild/buildhome/inactive_packages
-rm "$BUILDLOCATION"/build/"$BUILDARCH"/srcbuild/buildhome/inactive_packages/*
+if [[ $(ls -A "$BUILDLOCATION"/build/"$BUILDARCH"/srcbuild/buildhome/inactive_packages/) ]]
+then
+  rm "$BUILDLOCATION"/build/"$BUILDARCH"/srcbuild/buildhome/inactive_packages/*
+fi
 find "$BUILDLOCATION"/build/"$BUILDARCH"/srcbuild/ -mindepth 1 -maxdepth 1 -type d -printf "%f\n" | grep -v ^buildoutput$ | grep -v ^buildhome$ | while read -r PACKAGE
 do
   touch "$BUILDLOCATION"/build/"$BUILDARCH"/srcbuild/buildhome/inactive_packages/"$PACKAGE"
